+++ gw_full_latest/scripts/start_gw.sh
@@ -35,5 +35,16 @@ then
 	node-red-start &
 fi
 
+#check if the gateway is an access point
+#if yes, then enable IP forwarding to give internet connectivity to connected devices, e.g. smartphone, tablets,...
+if [ -f /etc/network/interfaces_not_ap ]
+then
+	sudo sh -c "echo 1 > /proc/sys/net/ipv4/ip_forward"
+	sudo sed -i 's/^#net.ipv4.ip_forward=1/net.ipv4.ip_forward=1/g' /etc/sysctl.conf
+	sudo iptables -t nat -A POSTROUTING -o eth0 -j MASQUERADE
+	sudo iptables -A FORWARD -i eth0 -o wlan0 -m state --state RELATED,ESTABLISHED -j ACCEPT
+	sudo iptables -A FORWARD -i wlan0 -o eth0 -j ACCEPT
+fi
+	
 #run the gateway
 python start_gw.py &